* Restructured microservices
* Cleaner Architecture
