+++ apps/auth/src/services/auth.service.ts
@@ -3,7 +3,7 @@ import { UserDocument } from '@app/common';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 import { Response } from 'express';
-import { TokenPayload } from './interfaces/token-payload.interface';
+import { TokenPayload } from '../interfaces/token-payload.interface';
 
 @Injectable()
 export class AuthService {
