+++ apps/auth/test/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { AuthModule } from './../src/auth.module';
+import { AuthModule } from '../src/modules/auth.module';
 
 describe('AuthController (e2e)', () => {
   let app: INestApplication;
