+++ apps/notifications/src/controllers/notifications.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, UsePipes, ValidationPipe } from '@nestjs/common';
-import { NotificationsService } from './notifications.service';
+import { NotificationsService } from '../services/notifications.service';
 import { EventPattern, Payload } from '@nestjs/microservices';
-import { NotifyEmailDto } from './dto/notify-email.dto';
+import { NotifyEmailDto } from '../dto/notify-email.dto';
 
 @Controller()
 export class NotificationsController {
