+++ apps/notifications/test/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import { NotificationsModule } from './../src/notifications.module';
+import { NotificationsModule } from '../src/modules/notifications.module';
 
 describe('NotificationsController (e2e)', () => {
   let app: INestApplication;
