+++ apps/payments/src/controllers/payments.controller.ts
@@ -1,12 +1,12 @@
 import { Controller, UsePipes, ValidationPipe } from '@nestjs/common';
-import { PaymentsService } from './payments.service';
+import { PaymentsService } from '../services/payments.service';
 import {
   Ctx,
   MessagePattern,
   Payload,
   RmqContext,
 } from '@nestjs/microservices';
-import { PaymentsCreateChargeDto } from './dto/payments-create-charge.dto';
+import { PaymentsCreateChargeDto } from '../dto/payments-create-charge.dto';
 
 @Controller()
 export class PaymentsController {
