+++ apps/reservations/src/repositories/reservations.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { AbstractRepository } from '@app/common';
-import { ReservationDocument } from './models/reservation.schema';
+import { ReservationDocument } from '../models/reservation.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
